Move cf_units_to_datetime
